requirements.txt
asgiref==3.8.1
astroid==2.4.2
colorama==0.4.6
Django==3.2.10
django-filter==24.2
djangorestframework==3.11.0
Faker==8.3.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pylint==2.5.3
python-dateutil==2.9.0.post0
pytz==2020.1
setuptools==69.5.1
six==1.16.0
sqlparse==0.3.1
text-unidecode==1.3
toml==0.10.1
tzdata==2024.1
validate-docbr==1.10.0
wrapt==1.12.1
